+++ docs/CHANGELOG.md
@@ -5,9 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-
 ### Added
 * Button to delete a specific work session
+
+### Fixed
+* When a work session is deleted, also the associated sprints are deleted
 
 ## 1.2.0-alpha - 2018-11-19
 ### Added